+++ data/types/base_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/tidepool-org/platform/app"
 	"github.com/tidepool-org/platform/data"
 	"github.com/tidepool-org/platform/data/types"
+	"github.com/tidepool-org/platform/pointer"
 )
 
 // TODO: Finish tests
@@ -29,7 +30,7 @@ var _ = Describe("Base", func() {
 				deviceID = app.NewID()
 				testBase.UserID = userID
 				testBase.DeviceID = &deviceID
-				testBase.Time = app.StringAsPointer("2016-09-06T13:45:58-07:00")
+				testBase.Time = pointer.String("2016-09-06T13:45:58-07:00")
 				testBase.Type = "testBase"
 			})
 
@@ -48,7 +49,7 @@ var _ = Describe("Base", func() {
 			})
 
 			It("returns error if device id is empty", func() {
-				testBase.DeviceID = app.StringAsPointer("")
+				testBase.DeviceID = pointer.String("")
 				identityFields, err := testBase.IdentityFields()
 				Expect(err).To(MatchError("base: device id is empty"))
 				Expect(identityFields).To(BeEmpty())
@@ -62,7 +63,7 @@ var _ = Describe("Base", func() {
 			})
 
 			It("returns error if time is empty", func() {
-				testBase.Time = app.StringAsPointer("")
+				testBase.Time = pointer.String("")
 				identityFields, err := testBase.IdentityFields()
 				Expect(err).To(MatchError("base: time is empty"))
 				Expect(identityFields).To(BeEmpty())